+++ client/plots/diffAnalysis/DifferentialAnalysis.ts
@@ -4,7 +4,8 @@ import { getCompInit, copyMerge, type RxComponent } from '#rx'
 import { PlotBase } from '../PlotBase'
 import { importPlot } from '../importPlot.js'
 import { Menu } from '#dom'
-import { GENE_EXPRESSION, termType2label, SINGLECELL_CELLTYPE, DNA_METHYLATION, PROTEOME_DAP } from '#shared/terms.js'
+import { GENE_EXPRESSION, SINGLECELL_CELLTYPE, DNA_METHYLATION, PROTEOME_DAP } from '#types'
+import { termType2label } from '#shared/terms.js'
 import type { DiffAnalysisDom, /*DiffAnalysisOpts,*/ DiffAnalysisPlotConfig } from './DiffAnalysisTypes'
 import { DiffAnalysisView } from './view/DiffAnalysisView'
 import { getDefaultVolcanoSettings, validateVolcanoSettings } from '../volcano/settings/defaults.ts'
