+++ client/plots/wsiviewer/viewModel/SessionWSImage.ts
@@ -1,6 +1,5 @@
-import type { Annotation, Prediction, TileSelection } from '@sjcrh/proteinpaint-types'
-import { WSImage } from '@sjcrh/proteinpaint-types'
-import { SelectionPrefixes, checkSelectionType, FlagStatusMessages, FlagStatus } from '#shared'
+import type { Annotation, Prediction, TileSelection } from '#types'
+import { WSImage, SelectionPrefixes, checkSelectionType, FlagStatusMessages, FlagStatus } from '#types'
 import type Settings from '#plots/wsiviewer/Settings.ts'
 export class SessionWSImage extends WSImage {
 	sessionsTileSelections?: TileSelection[]
