+++ shared/types/src/index.ts
@@ -5,6 +5,8 @@ export * from './termsetting.ts'
 export * from './filter.ts'
 export * from './termdb.matrix.ts'
 export * from './routes/routeApi.ts'
+export * from './routes/alphaGenome.ts'
+export * from './routes/alphaGenomeTypes.ts'
 export * from './routes/aiProjectAdmin.ts'
 export * from './routes/aiProjectSelectedWSImages.ts'
 export * from './routes/aiProjectTrainModel.ts'
@@ -74,6 +76,7 @@ export * from './routes/wsimages.ts'
 export * from './routes/wsisamples.ts'
 export * from './terms/categorical.ts'
 export * from './terms/condition.ts'
+export * from './terms/constants.ts'
 export * from './terms/date.ts'
 export * from './terms/numeric.ts'
 export * from './terms/geneVariant.ts'
