+++ shared/types/tsconfig.json
@@ -15,6 +15,7 @@
   "exclude": [
     "./node_modules",
     "./*.ts",
-    "./dist"
+    "./dist",
+  	"./package"
   ]
 }
