--- shared/utils/devTs.ts
@@ -1,4 +1,0 @@
-// please list in alphanumeric order for readability
-// use .ts file extension for all code files below
-export * from './constants/AiHisto.js'
-export * from './constants/terms.js'
