+++ shared/utils/package.json
@@ -21,17 +21,16 @@
     "./*": {
       "sjpp/dev": "./src/*",
       "default": "./dist/src/*"
-    },
-    "./devTs": "./devTs.ts"
+    }
   },
   "scripts": {
-    "ts2js": "esbuild src/*.* constants/*.ts --platform=node --outdir=dist --format=esm --sourcemap",
-    "build": "rm -rf dist && npm run ts2js && tsc --sourcemap --skipLibCheck",
+    "ts2js": "esbuild src/*.* src/**/*.* --platform=node --outdir=dist/src --format=esm --sourcemap",
+    "build": "rm -rf dist && npm run ts2js && rm -rf dist/**/*.spec.* dist/**/test && tsc --sourcemap --skipLibCheck",
     "prepack": "npm run build",
     "dev0": "rm -rf dist && 'npx tsc --watch --skipLibCheck --preserveWatchOutput'",
-    "dev": "echo '--- now using raw shared/utils/src exports under conditions=sjpp/dev ---'",
+    "dev": "echo '--- now using raw shared/utils/src exports under conditions=sjpp/dev ---\n'",
     "pretest": "mkdir -p test && node emitImports > test/internals-test.ts",
-    "test": "tsx test/internals-test.ts",
+    "test": "tsx --conditions=sjpp/dev test/internals-test.ts",
     "test-x": "ls src/test/*.spec* | xargs -I % bash -c '{ tsx %; sleep 0.001; }'",
     "spec:coverage": "node test/relevant.js"
   },
